# contracts/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ampara Animal — contrato do Ciclo 1",
    "version": "0.3.0",
    "description": "Contrato do Ciclo 1 (Sprint 3) e do prontuário do Ciclo 2. Endpoints marcados x-situacao=implementado existem no código e são conferidos por scripts/validar-openapi.cjs. Todas as rotas exigem login (Authorization: Bearer <access token do Supabase Auth>), exceto /api/health e /api/health/ready. IDs e dados dos exemplos são sintéticos. Autorização: além do login, a conta precisa estar liberada na API (tabela Usuario); conta não liberada ou desativada recebe 403. ADMIN acessa todas as unidades; OPERADOR só as vinculadas (papéis provisórios até aprovação da Ampara). O escopo é decidido no servidor: filtros enviados pelo cliente nunca ampliam o acesso."
  },
  "paths": {
...
    "/api/v1/species": {
      "get": {
        "summary": "Listar espécies",
        "description": "Implementado. Lista espécies ativas; o campo silvestre indica identificação por nome.",
        "x-situacao": "implementado",
        "security": [
          {
            "bearerSupabase": []
          }
        ],
        "responses": {
          "200": {
            "description": "Resultado esperado quando implementado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginaCatalogo"
                }
              }
            }
          },
          "400": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "401": {
            "description": "Sem token, token inválido ou vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "403": {
            "description": "Conta não liberada/desativada, ou registro fora das unidades do usuário",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "404": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "409": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "422": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "429": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "500": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "503": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "pagina",
            "in": "query",
            "required": false,
            "description": "Filtro ou controle da requisição",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limite",
            "in": "query",
            "required": false,
            "description": "Filtro ou controle da requisição",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ]
      }
    },
    "/api/v1/breeds": {
      "get": {
        "summary": "Listar raças",
        "description": "Implementado. Lista raças de uma espécie (especieId obrigatório).",
        "x-situacao": "implementado",
        "security": [
          {
            "bearerSupabase": []
          }
        ],
        "responses": {
          "200": {
            "description": "Resultado esperado quando implementado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginaCatalogo"
                }
              }
            }
          },
          "400": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "401": {
            "description": "Sem token, token inválido ou vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "403": {
            "description": "Conta não liberada/desativada, ou registro fora das unidades do usuário",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "404": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "409": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "422": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "429": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "500": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "503": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "pagina",
            "in": "query",
            "required": false,
            "description": "Filtro ou controle da requisição",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limite",
            "in": "query",
            "required": false,
            "description": "Filtro ou controle da requisição",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "especieId",
            "in": "query",
            "required": true,
            "description": "Filtro ou controle da requisição",
            "schema": {
              "type": "string",
              "description": "Identificador UUID",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/api/v1/units": {
      "get": {
        "summary": "Listar unidades",
        "description": "Implementado. Lista só as unidades em que o usuário pode trabalhar (todas, para ADMIN).",
        "x-situacao": "implementado",
        "security": [
          {
            "bearerSupabase": []
          }
        ],
        "responses": {
          "200": {
            "description": "Resultado esperado quando implementado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginaCatalogo"
                }
              }
            }
          },
          "400": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "401": {
            "description": "Sem token, token inválido ou vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "403": {
            "description": "Conta não liberada/desativada, ou registro fora das unidades do usuário",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "404": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "409": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "422": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "429": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "500": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "503": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "pagina",
            "in": "query",
            "required": false,
            "description": "Filtro ou controle da requisição",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limite",
            "in": "query",
            "required": false,
            "description": "Filtro ou controle da requisição",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ]
      }
    },
    "/api/v1/locations": {
      "get": {
        "summary": "Listar localizações",
        "description": "Implementado. Lista localizações de uma unidade (unidadeId obrigatório); unidade fora do escopo do usuário: 403.",
        "x-situacao": "implementado",
        "security": [
          {
            "bearerSupabase": []
          }
        ],
        "responses": {
          "200": {
            "description": "Resultado esperado quando implementado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginaCatalogo"
                }
              }
            }
          },
          "400": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "401": {
            "description": "Sem token, token inválido ou vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "403": {
            "description": "Conta não liberada/desativada, ou registro fora das unidades do usuário",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "404": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "409": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "422": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "429": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "500": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "503": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "pagina",
            "in": "query",
            "required": false,
            "description": "Filtro ou controle da requisição",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limite",
            "in": "query",
            "required": false,
            "description": "Filtro ou controle da requisição",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "unidadeId",
            "in": "query",
            "required": true,
            "description": "Filtro ou controle da requisição",
            "schema": {
              "type": "string",
              "description": "Identificador UUID",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/api/v1/responsibles": {
      "get": {
        "summary": "Listar responsáveis",
        "description": "Implementado. Lista responsáveis ativos e não arquivados.",
        "x-situacao": "implementado",
        "security": [
          {
            "bearerSupabase": []
          }
        ],
        "responses": {
          "200": {
            "description": "Resultado esperado quando implementado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginaCatalogo"
                }
              }
            }
          },
          "400": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "401": {
            "description": "Sem token, token inválido ou vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "403": {
            "description": "Conta não liberada/desativada, ou registro fora das unidades do usuário",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "404": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "409": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "422": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "429": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "500": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "503": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "pagina",
            "in": "query",
            "required": false,
            "description": "Filtro ou controle da requisição",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limite",
            "in": "query",
            "required": false,
            "description": "Filtro ou controle da requisição",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ]
      },
      "post": {
        "summary": "Cadastrar responsável",
        "description": "Implementado. Cadastra o tutor ou quem entregou o animal, para uso em responsavelId.",
        "x-situacao": "implementado",
        "security": [
          {
            "bearerSupabase": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CriarResponsavel"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Responsável cadastrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Responsavel"
                }
              }
            }
          },
          "400": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "401": {
            "description": "Sem token, token inválido ou vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "403": {
            "description": "Conta não liberada/desativada, ou registro fora das unidades do usuário",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "429": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "500": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fronts": {
      "get": {
        "summary": "Listar frentes",
        "description": "Implementado. Lista fixa das três frentes de atuação (CCPA, CasAdote, CED).",
        "x-situacao": "implementado",
        "security": [
          {
            "bearerSupabase": []
          }
        ],
        "responses": {
          "200": {
            "description": "Resultado esperado quando implementado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginaFrentes"
                }
              }
            }
          },
          "400": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "401": {
            "description": "Sem token, token inválido ou vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "403": {
            "description": "Conta não liberada/desativada, ou registro fora das unidades do usuário",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "404": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "409": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "422": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "429": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "500": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "503": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "pagina",
            "in": "query",
            "required": false,
            "description": "Filtro ou controle da requisição",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limite",
            "in": "query",
            "required": false,
            "description": "Filtro ou controle da requisição",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Erro": {
        "type": "object",
        "required": [
          "statusCode",
          "mensagem",
          "caminho",
          "timestamp"
        ],
        "properties": {
          "statusCode": {
            "type": "integer"
          },
          "mensagem": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ]
          },
          "caminho": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "description": "Instante ISO-8601",
            "format": "date-time"
          },
          "detalhes": {
            "type": "object",
            "additionalProperties": true,
            "description": "Opcional, só em alguns 4xx. Ex.: 409 de microchip duplicado traz animalExistenteId e identificadorPublico para o frontend oferecer a transferência."
          }
        },
        "example": {
          "statusCode": 409,
          "mensagem": "Versão desatualizada",
          "caminho": "/api/v1/animals/56b9f220-3c58-4adc-9dfa-102b90720e12",
          "timestamp": "2026-09-13T12:00:00Z"
        }
      },
...
      "Catalogo": {
        "type": "object",
        "required": [
          "id",
          "nome"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Identificador UUID",
            "format": "uuid"
          },
          "nome": {
            "type": "string",
            "description": "Nome de catálogo"
          },
          "especieId": {
            "type": "string",
            "description": "Identificador UUID",
            "format": "uuid"
          },
          "unidadeId": {
            "type": "string",
            "description": "Identificador UUID",
            "format": "uuid"
          },
          "silvestre": {
            "type": "boolean",
            "description": "Presente apenas em espécies"
          }
        },
        "example": {
          "id": "56b9f220-3c58-4adc-9dfa-102b90720e12",
          "nome": "Exemplo de desenvolvimento"
        }
      },
      "Frente": {
        "type": "object",
        "required": [
          "codigo",
          "nome"
        ],
        "properties": {
          "codigo": {
            "type": "string",
            "description": "Classificação atual; demais nomes aguardam validação",
            "enum": [
              "CCPA",
              "CASADOTE",
              "CED"
            ]
          },
          "nome": {
            "type": "string",
            "description": "Nome de exibição"
          }
        },
        "example": {
          "codigo": "CASADOTE",
          "nome": "CasAdote"
        }
      },
...
      "PaginaCatalogo": {
        "type": "object",
        "required": [
          "itens",
          "pagina",
          "limite",
          "total"
        ],
        "properties": {
          "itens": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Catalogo"
            }
          },
          "pagina": {
            "type": "integer",
            "minimum": 1
          },
          "limite": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          },
          "total": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "PaginaFrentes": {
        "type": "object",
        "required": [
          "itens",
          "pagina",
          "limite",
          "total"
        ],
        "properties": {
          "itens": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Frente"
            }
          },
          "pagina": {
            "type": "integer",
            "minimum": 1
          },
          "limite": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          },
          "total": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
      "CriarResponsavel": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "nome"
        ],
        "properties": {
          "nome": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120
          },
          "endereco": {
            "type": "string",
            "maxLength": 300,
            "nullable": true
          },
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 160,
            "nullable": true
          },
          "telefone": {
            "type": "string",
            "maxLength": 30,
            "nullable": true,
            "example": "(11) 91234-5678"
          }
        },
        "example": {
          "nome": "Maria Exemplo",
          "telefone": "(11) 91234-5678"
        }
      },
      "Responsavel": {
        "type": "object",
        "required": [
          "id",
          "nome"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nome": {
            "type": "string"
          },
          "endereco": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "telefone": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerSupabase": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token emitido pelo Supabase Auth do projeto (o mesmo login do frontend). Validado no servidor pela assinatura (JWKS), emissor e audiência \"authenticated\". Sem token ou com token inválido/vencido: 401."
      }
    }
  }
}